+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#686](https://github.com/NibiruChain/nibiru/pull/686) Add changelog enforcer to github actions.
 * [#681](https://github.com/NibiruChain/nibiru/pull/681) Remove automatic release and leave integration tests when merge into master.
 * [#684](https://github.com/NibiruChain/nibiru/pull/684) Reorganize PerpKeeper methods.
+* [#690](https://github.com/NibiruChain/nibiru/pull/690) Call `closePositionEntirely` from `ClosePosition`.
 * [#689](https://github.com/NibiruChain/nibiru/pull/689) Apply funding rate calculation 48 times per day.
 
 ### API Breaking